# target/classes/openapi.yaml
openapi: 3.0.0
info:
  title: Get Items in Custom Carousel
  version: 1.0.0
servers:
- url: /api
paths:
...
  /blogs/{title}:
    delete:
      parameters:
      - description: title of the blog
        explode: false
        in: path
        name: title
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Blog deleted successfully
        "400":
          description: Invalid request
      summary: Delete a specific blog by its ID
      x-accepts: application/json
    get:
      parameters:
      - description: title of the blog
        explode: false
        in: path
        name: title
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blog'
          description: Blog retrieved successfully
        "404":
          description: Blog not found
      summary: Return a specific blog by its ID
      x-accepts: application/json
    put:
      parameters:
      - description: title of the blog
        explode: false
        in: path
        name: title
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Blog'
        required: true
      responses:
        "200":
          description: Blog updated successfully
        "400":
          description: Invalid request payload
      summary: Update a specific blog by its ID
      x-content-type: application/json
      x-accepts: application/json
...
  /categories/{title}:
    delete:
      parameters:
      - description: title of the category
        explode: false
        in: path
        name: title
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Category deleted successfully
        "400":
          description: Invalid request
      summary: Delete a specific category by its title
      x-accepts: application/json
    get:
      parameters:
      - description: title of the category
        explode: false
        in: path
        name: title
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
          description: Category return successfully
        "404":
          description: Category not found
      summary: Retrieve a specific category (Search by key of title)
      x-accepts: application/json
    post:
      parameters:
      - description: title of the category
        explode: false
        in: path
        name: title
        required: true
        schema:
          type: string
        style: simple
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
          description: Category created successfully
        "400":
          description: Invalid request payload
      summary: Create a new category with a custom key
      x-content-type: application/json
      x-accepts: application/json
...
components:
  schemas:
    Blog:
      example:
        image: https://example.com/aeiou
        link: https://example.com/aeiou
        title: title
      properties:
        title:
          description: The title of the blog
          type: string
        image:
          description: The URL of the blog image
          format: url
          type: string
        link:
          description: The URL of the blog link
          format: url
          type: string
      type: object
...
    Category:
      example:
        title: title
        key: key
      properties:
        title:
          description: The title of the category
          type: string
        key:
          description: The key of the category
          type: string
      type: object
...